+++ src/apps/toolbar/modules/Notifications/domain.ts
@@ -1,8 +1,8 @@
 export interface AppNotification {
   id: number;
-  app_name: string;
-  app_description: string;
-  app_umid: string;
+  appName: string;
+  appDescription: string;
+  appUmid: string;
   body: string[];
   date: number;
 }